+++ src/Support/Adapters/ExportService.php
@@ -11,17 +11,15 @@ abstract class ExportService
      * Provide a view to build the export from.
      *
      * @param View $view
-     * @param string|null $uploadPath
      * @return Renderer
      */
-    abstract public static function fromView(View $view, string $uploadPath = null): Renderer;
+    abstract public static function fromView(View $view): Renderer;
 
     /**
      * Provide a view to build the export from.
      *
      * @param AbstractViewModel $viewModel
-     * @param string|null $uploadPath
      * @return Renderer
      */
-    abstract public static function fromViewModel(AbstractViewModel $viewModel, string $uploadPath = null): Renderer;
+    abstract public static function fromViewModel(AbstractViewModel $viewModel): Renderer;
 }